requirements.txt
requests>=2.31.0
openai>=1.12.0
httpx>=0.27.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
tweepy>=4.14.0
fastapi>=0.109.0
uvicorn>=0.27.0
